# apps/console/openapi/p1.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Operations Performance API",
  "description": "Analytics API over the Northstar Manufacturing procurement process. Read-only except the admin-only intervention ledger (POST /interventions). Wraps the same analytics/ML logic used by the local pipeline and dashboard, so there's one source of truth for every number -- including for operations-assistant, which consumes these endpoints as agent tools rather than reimplementing the logic.",
  "version": "0.1.0"
 },
 "paths": {
  "/v1/mlops/registry": {
   "get": {
    "tags": [
     "v1"
    ],
    "summary": "V1 Mlops Registry",
    "description": "Champion/challenger state, version list and the latest promotion/rollback events (src/mlops). Read-only; creates nothing.",
    "operationId": "v1_mlops_registry_v1_mlops_registry_get",
    "parameters": [
     {
      "name": "name",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "pattern": "^[A-Za-z0-9_-]{1,64}$",
       "default": "sla_risk",
       "title": "Name"
      }
     },
     {
      "name": "events",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 100,
       "minimum": 0,
       "default": 15,
       "title": "Events"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/RegistryStatus"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/overview": {
   "get": {
    "tags": [
     "v1"
    ],
    "summary": "V1 Overview",
    "operationId": "v1_overview_v1_overview_get",
    "parameters": [
     {
      "name": "as_of",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "description": "Replay clock (ISO date/time). Open cases and their features use only events up to this instant.",
       "default": "2018-04-16",
       "title": "As Of"
      },
      "description": "Replay clock (ISO date/time). Open cases and their features use only events up to this instant."
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/queue": {
   "get": {
    "tags": [
     "v1"
    ],
    "summary": "V1 Queue",
    "operationId": "v1_queue_v1_queue_get",
    "parameters": [
     {
      "name": "as_of",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "description": "Replay clock (ISO date/time). Open cases and their features use only events up to this instant.",
       "default": "2018-04-16",
       "title": "As Of"
      },
      "description": "Replay clock (ISO date/time). Open cases and their features use only events up to this instant."
     },
     {
      "name": "min_value",
      "in": "query",
      "required": false,
      "schema": {
       "type": "number",
       "default": 0.0,
       "title": "Min Value"
      }
     },
     {
      "name": "supplier",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Supplier"
      }
     },
     {
      "name": "stage",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Stage"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 500,
       "minimum": 1,
       "default": 100,
       "title": "Limit"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/cases/{case_id}": {
   "get": {
    "tags": [
     "v1"
    ],
    "summary": "V1 Case",
    "operationId": "v1_case_v1_cases__case_id__get",
    "parameters": [
     {
      "name": "case_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Case Id"
      }
     },
     {
      "name": "as_of",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "description": "Replay clock (ISO date/time). Open cases and their features use only events up to this instant.",
       "default": "2018-04-16",
       "title": "As Of"
      },
      "description": "Replay clock (ISO date/time). Open cases and their features use only events up to this instant."
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/suppliers": {
   "get": {
    "tags": [
     "v1"
    ],
    "summary": "V1 Suppliers",
    "operationId": "v1_suppliers_v1_suppliers_get",
    "parameters": [
     {
      "name": "as_of",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "description": "Replay clock (ISO date/time). Open cases and their features use only events up to this instant.",
       "default": "2018-04-16",
       "title": "As Of"
      },
      "description": "Replay clock (ISO date/time). Open cases and their features use only events up to this instant."
     },
     {
      "name": "sort",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "default": "ci_lower",
       "title": "Sort"
      }
     },
     {
      "name": "min_n",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 1,
       "default": 20,
       "title": "Min N"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/suppliers/{supplier_id}": {
   "get": {
    "tags": [
     "v1"
    ],
    "summary": "V1 Supplier",
    "operationId": "v1_supplier_v1_suppliers__supplier_id__get",
    "parameters": [
     {
      "name": "supplier_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Supplier Id"
      }
     },
     {
      "name": "as_of",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "description": "Replay clock (ISO date/time). Open cases and their features use only events up to this instant.",
       "default": "2018-04-16",
       "title": "As Of"
      },
      "description": "Replay clock (ISO date/time). Open cases and their features use only events up to this instant."
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/process/flow": {
   "get": {
    "tags": [
     "v1"
    ],
    "summary": "V1 Flow",
    "operationId": "v1_flow_v1_process_flow_get",
    "parameters": [
     {
      "name": "as_of",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "description": "Replay clock (ISO date/time). Open cases and their features use only events up to this instant.",
       "default": "2018-04-16",
       "title": "As Of"
      },
      "description": "Replay clock (ISO date/time). Open cases and their features use only events up to this instant."
     },
     {
      "name": "metric",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "pattern": "^(median|p75|p90|breach_contribution)$",
       "default": "median",
       "title": "Metric"
      }
     },
     {
      "name": "group",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "pattern": "^(stage|activity)$",
       "default": "stage",
       "title": "Group"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/process/variants": {
   "get": {
    "tags": [
     "v1"
    ],
    "summary": "V1 Variants",
    "operationId": "v1_variants_v1_process_variants_get",
    "parameters": [
     {
      "name": "as_of",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "description": "Replay clock (ISO date/time). Open cases and their features use only events up to this instant.",
       "default": "2018-04-16",
       "title": "As Of"
      },
      "description": "Replay clock (ISO date/time). Open cases and their features use only events up to this instant."
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/risk-map": {
   "get": {
    "tags": [
     "v1"
    ],
    "summary": "V1 Risk Map",
    "operationId": "v1_risk_map_v1_risk_map_get",
    "parameters": [
     {
      "name": "as_of",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "description": "Replay clock (ISO date/time). Open cases and their features use only events up to this instant.",
       "default": "2018-04-16",
       "title": "As Of"
      },
      "description": "Replay clock (ISO date/time). Open cases and their features use only events up to this instant."
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/finance/controls": {
   "get": {
    "tags": [
     "v1"
    ],
    "summary": "V1 Fin Controls",
    "operationId": "v1_fin_controls_v1_finance_controls_get",
    "parameters": [
     {
      "name": "as_of",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "description": "Replay clock (ISO date/time). Open cases and their features use only events up to this instant.",
       "default": "2018-04-16",
       "title": "As Of"
      },
      "description": "Replay clock (ISO date/time). Open cases and their features use only events up to this instant."
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/finance/exceptions": {
   "get": {
    "tags": [
     "v1"
    ],
    "summary": "V1 Fin Exc",
    "operationId": "v1_fin_exc_v1_finance_exceptions_get",
    "parameters": [
     {
      "name": "control",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Control"
      }
     },
     {
      "name": "min_exposure",
      "in": "query",
      "required": false,
      "schema": {
       "type": "number",
       "default": 0.0,
       "title": "Min Exposure"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/v1/lineage": {
   "get": {
    "tags": [
     "v1"
    ],
    "summary": "V1 Lineage",
    "operationId": "v1_lineage_v1_lineage_get",
    "parameters": [
     {
      "name": "metric",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Metric"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/v1/briefing": {
   "get": {
    "tags": [
     "v1"
    ],
    "summary": "V1 Briefing",
    "operationId": "v1_briefing_v1_briefing_get",
    "parameters": [
     {
      "name": "as_of",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "description": "Replay clock (ISO date/time). Open cases and their features use only events up to this instant.",
       "default": "2018-04-16",
       "title": "As Of"
      },
      "description": "Replay clock (ISO date/time). Open cases and their features use only events up to this instant."
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/search": {
   "get": {
    "tags": [
     "v1"
    ],
    "summary": "V1 Search",
    "operationId": "v1_search_v1_search_get",
    "parameters": [
     {
      "name": "q",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "minLength": 1,
       "title": "Q"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 50,
       "minimum": 1,
       "default": 20,
       "title": "Limit"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/interventions/{intervention_id}/outcome": {
   "patch": {
    "tags": [
     "v1"
    ],
    "summary": "V1 Record Outcome",
    "description": "Record whether the case breached after an intervention (admin key required). Outcomes on holdout rows are what\nturn the assumed effects into measured uplift; until some exist, /v1/interventions/roi reports only simulated results.",
    "operationId": "v1_record_outcome_v1_interventions__intervention_id__outcome_patch",
    "parameters": [
     {
      "name": "intervention_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Intervention Id"
      }
     },
     {
      "name": "x-api-key",
      "in": "header",
      "required": false,
      "schema": {
       "type": "string",
       "title": "X-Api-Key"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/OutcomeBody"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  }
 },
 "components": {
  "schemas": {
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "OutcomeBody": {
    "properties": {
     "breached_after": {
      "type": "boolean",
      "title": "Breached After"
     },
     "notes": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 500
       },
       {
        "type": "null"
       }
      ],
      "title": "Notes"
     }
    },
    "type": "object",
    "required": [
     "breached_after"
    ],
    "title": "OutcomeBody"
   },
...
   "RegistryEvent": {
    "properties": {
     "at": {
      "type": "string",
      "title": "At"
     },
     "event": {
      "type": "string",
      "title": "Event"
     },
     "detail": {
      "additionalProperties": true,
      "type": "object",
      "title": "Detail"
     }
    },
    "type": "object",
    "required": [
     "at",
     "event"
    ],
    "title": "RegistryEvent"
   },
   "RegistryModel": {
    "properties": {
     "version": {
      "type": "string",
      "title": "Version"
     },
     "registered_at": {
      "type": "string",
      "title": "Registered At"
     },
     "sha256": {
      "type": "string",
      "title": "Sha256"
     },
     "data_fingerprint": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Data Fingerprint"
     },
     "metrics": {
      "additionalProperties": true,
      "type": "object",
      "title": "Metrics"
     },
     "notes": {
      "type": "string",
      "title": "Notes",
      "default": ""
     }
    },
    "type": "object",
    "required": [
     "version",
     "registered_at",
     "sha256",
     "metrics"
    ],
    "title": "RegistryModel"
   },
   "RegistryStatus": {
    "properties": {
     "available": {
      "type": "boolean",
      "title": "Available"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "champion": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/RegistryModel"
       },
       {
        "type": "null"
       }
      ]
     },
     "challenger": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/RegistryModel"
       },
       {
        "type": "null"
       }
      ]
     },
     "versions": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Versions"
     },
     "champion_hash_verified": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Champion Hash Verified"
     },
     "events": {
      "items": {
       "$ref": "#/components/schemas/RegistryEvent"
      },
      "type": "array",
      "title": "Events"
     },
     "note": {
      "type": "string",
      "title": "Note",
      "default": ""
     }
    },
    "type": "object",
    "required": [
     "available",
     "name"
    ],
    "title": "RegistryStatus",
    "description": "Model lifecycle state read straight from the file registry (no database). `available` is False until a model has been registered."
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  }
 }
}